完善代码结构，优化细节。 现在插件如果获取不到taskId会提出错误了。
